Add files via upload
AIM -The Aim of this project is to make valuable predictions on  year over year growth which is very essential in investment banking considering India, as investors usually want to see your year-over-year numbers before supplying you with business capital as well as YOY is a great statistic to use when you want to control the effects of volatility when comparing companies or economies.
CONTENT - In order to make predictions on year over year growth we have gathered data from Reserve bank of India. Followings parameters have been taken into account to make predictions on year over year growth including Fortnight Ended , Non-food Bank Credit , 	Investments in Commercial Paper ,Investments in shares,	Investments in Bonds/Debentures,	Total Non-SLR Investments and	Adjusted Non-food Bank Credit.

Description of Columns of dataset :

1-Non Food Bank Credit-
Banks provide credit or loans to various organizations and when this loans are given to Food Corporation (FCI)or for promotion of food safety etc they are known as Food credit and loans other that related to food are called Non food bank credit.Food Credit is related to bank finance given to FCI for procurement of food grains. Other than this all are non-food bank credit.
Banks non-food bank credit growth decelerated to 7.2 per cent in November 2019 from 13.8 per cent in November 2018, on a year-on-year (y-o-y) basis,according to the latest data from the Reserve Bank of India. Credit to agriculture & allied activities slowed down to 6.5 per cent in November 2019 from 7.7 per cent in November 2018. Credit growth to industry decelerated to 2.4 per cent in November 2019 from 4.0 per cent in November 2018. Within industry, credit growth to 'food processing', 'textiles', 'gems & jewellery', 'chemicals & chemicals products', 'infrastructure' and 'construction'

2-Investments in Commercial Paper-
Commercial paper (CP) is a short-term debt instrument issued by companies to raise funds for a time period of up to one year. Introduced in 1990, it is an unsecured money market instrument issued in the form of a promissory note.Any corporate with a high rating can diversify their sources of short-term borrowings using CP. Investors get an additional instrument to park money. This instrument is typically issued by large banks or corporations to cover short-term receivables and meet short-term financial obligations, such as funding for a new project.

Trading in Commercial Paper-
It is possible for small retail investors to purchase commercial paper, although there are several restrictions that make it more difficult. Most commercial paper is sold and resold to institutional investors, such as large financial institutions, hedge funds, and multinational corporations. A retail investor would need access to very large amounts of capital to buy and own commercial paper; otherwise, indirect investment is possible through mutual funds, exchange-traded funds (ETFs) or a money market account administered and held at a depository institution.Factors such as regulatory costs, the scale of investable capital, and physical access to the capital markets can make it very difficult for an individual or retail investors to buy and own commercial paper.For example, commercial paper is typically sold in round lots totaling $100,000. This threshold in itself makes buying commercial paper generally exclusive to institutional investors and wealthy individuals. Further, broker-dealers issuing commercial paper on behalf of a client have pre-existing relationships with institutional buyers that make the market efficient through large purchases of primary offerings. They would not be likely to look to individual investors as a source of capital to fund the transaction.

3-Investments in Shares-
Investing in shares means buying and keeping them for a while in order to make money.There are two ways of getting money from shares of a company.If the company grows and becomes more valuable, the share is worth more – so your investment is worth more too.Some shares pay you part of the company’s profits each year, called a dividend.If you buy shares in larger, long-established companies you’ll probably get dividends, but you might not get rapid growth.Shares that pay regular dividends are good for getting an income or the dividends can be reinvested to grow your capital.Dividend income is taxed at a different rate from savings interest.Smaller companies often don’t pay dividends. They might have more chance to grow rapidly, but can be more risky.

4-Investments in Bonds-
Investment Bonds are debt instruments in which the authorized issuer owes the bond holders a debt. Depending on the terms of the type of bonds, the authorized issuer is obliged to pay interest and/or repay the principal at a later date upon maturity. In simpler terms, a bond is a formal contract to repay borrowed money with an interest at fixed intervals. Investment bonds are a way to raise money. When you purchase any type of bond (government, convertible, callable, etc.), you are lending money to the issuer which may be a corporation, the government, a federal agency or any other entity. In return, the issuer promises to pay a specified rate of interest during the life of the bond. The issuer also repays the face value of the bond when upon maturity of the term.Arguably, investing in bonds is simpler than doing so in equities, because there is a defined maturity. That is, the issuer of the debt instrument commits to pay back the invested amount on a specified date. In equities, there is no maturity, so the issuer would not repay you. Therefore, you have to sell it in the secondary market when you need to. The selling price of your equity shares depends on someone else, i.e., the purchaser at that point of time, over which you have no control.

5-Non SLR-Investments-
Besides giving loans to businesses and individuals, RBI has also allowed banks to invest in various capital market instruments such as stocks and bonds issued by public and private sector companies and commercial papers. In addition, banks are also allowed to invest in various mutual fund schemes. Unlike SLR investments, there is no compulsion on banks to invest in these instruments. Investments are entirely guided by commercial considerations and many such investments are in accordance with the prescribed guidelines.RBI treats both loans extended by commercial banks and the non-SLR investments as a resource flow to the commercial sector. Hence, it includes both while making credit projections it is comfortable with to achieve the targeted economic growth at the time of the monetary policy formulation during the beginning of the fiscal year.
6-Adjusted Non-Food Bank Credit-
Adjusted non-food bank credit includes non-food bank credit and total non-statutory liquidity ratio (SLR) investments of banks in commercial papers, shares and bonds/debentures.
7-Y-o-Y Growth in (7) (%)-
Year-Over-Year (YOY) is a frequently used financial comparison for comparing two or more measurable events on an annualized basis.Looking at YOY performance allows for gauging if a company's financial performance is improving, static, or worsening. For example, in financial reports, you may read that a particular business reported its revenues increased for the third quarter, on a YOY basis, for the last three years.YOY comparisons are a popular and effective way to evaluate the financial performance of a company and the performance of investments. Any measurable event that repeats annually can be compared on a YOY basis. Common YOY comparisons include annual, quarterly, and monthly performance.The period is for a month or quarter basis. The year-over-year growth rate calculates the percentage change during the past twelve months. Year-over-year (YOY) is an effective way of looking at growth for two reasons.If sales typically rise 35 percent this month, then your revenue is down year-over-year.